move usePositions to context
